refactor(server): move encryption to infrastructure ownership
