+++ server/alembic/versions/f8a9b0c1d2e3_auth_identity_multiple_provider_accounts.py
@@ -17,7 +17,7 @@ import sqlalchemy as sa
 
 from alembic import op
 from proliferate.config import settings
-from proliferate.utils.crypto import encrypt_text
+from proliferate.lib.infra.encryption.fernet import encrypt_text
 
 # revision identifiers, used by Alembic.
 revision: str = "f8a9b0c1d2e3"
