+++ server/proliferate/db/store/agent_gateway/api_keys.py
@@ -18,7 +18,8 @@ from proliferate.constants.agent_gateway import (
 from proliferate.db.models.cloud.agent_gateway import AgentApiKey
 from proliferate.db.store.agent_gateway.mappers import api_key_record
 from proliferate.db.store.agent_gateway.records import AgentApiKeyRecord
-from proliferate.utils.crypto import decrypt_json, decrypt_text, encrypt_json, encrypt_text
+from proliferate.lib.infra.encryption.fernet import decrypt_text, encrypt_text
+from proliferate.lib.infra.encryption.json import decrypt_json, encrypt_json
 from proliferate.utils.time import utcnow
