+++ server/proliferate/db/store/agent_gateway/enrollment_keys.py
@@ -18,7 +18,7 @@ from proliferate.constants.agent_gateway import AGENT_GATEWAY_CIPHERTEXT_KEY_ID
 from proliferate.db.models.cloud.agent_gateway import AgentGatewayEnrollmentKey
 from proliferate.db.store.agent_gateway.mappers import enrollment_key_record
 from proliferate.db.store.agent_gateway.records import AgentGatewayEnrollmentKeyRecord
-from proliferate.utils.crypto import decrypt_text, encrypt_text
+from proliferate.lib.infra.encryption.fernet import decrypt_text, encrypt_text
 from proliferate.utils.time import utcnow
