+++ server/proliferate/db/store/cloud_secrets.py
@@ -17,7 +17,7 @@ from proliferate.db.models.cloud.secrets import (
     CloudSecretFile,
     CloudSecretSet,
 )
-from proliferate.utils.crypto import decrypt_text, encrypt_text
+from proliferate.lib.infra.encryption.fernet import decrypt_text, encrypt_text
 from proliferate.utils.time import utcnow
