+++ server/proliferate/server/cloud/materialization/sandbox_io/connect.py
@@ -25,6 +25,7 @@ from proliferate.integrations.sandbox import (
     SandboxProviderUnavailableError,
     get_sandbox_provider,
 )
+from proliferate.lib.infra.encryption.fernet import decrypt_text, encrypt_text
 from proliferate.server.billing.authorization import assert_cloud_sandbox_resume_allowed
 from proliferate.server.billing.runtime_usage import (
     converge_cloud_sandbox_provider_usage,
@@ -52,7 +53,6 @@ from proliferate.server.cloud.runtime.liveness_health import (
     verify_runtime_auth_enforced,
     wait_for_runtime_health,
 )
-from proliferate.utils.crypto import decrypt_text, encrypt_text
 from proliferate.utils.time import utcnow
 
 logger = logging.getLogger("proliferate.cloud.materialization.connect")
