+++ server/tests/integration/test_cloud_sandbox_cold_access_repair.py
@@ -36,7 +36,7 @@ from proliferate.server.cloud.errors import CloudApiError
 from proliferate.server.cloud.gateway import service as gateway_service
 from proliferate.server.cloud.materialization import operation, runner
 from proliferate.server.cloud.materialization import service as materialization_service
-from proliferate.utils.crypto import encrypt_text
+from proliferate.lib.infra.encryption.fernet import encrypt_text
 from proliferate.utils.time import utcnow
